Move check_figures_equal out of decorators.py and back into testing.py
