+++ pygmt/helpers/__init__.py
@@ -1,7 +1,7 @@
 """
 Functions, classes, decorators, and context managers to help wrap GMT modules.
 """
-from .decorators import check_figures_equal, fmt_docstring, kwargs_to_strings, use_alias
+from .decorators import fmt_docstring, use_alias, kwargs_to_strings
 from .tempfile import GMTTempFile, unique_name
 from .utils import (
     data_kind,
